SCHEMA CHANGE: add HostCategory.always_up2date
There are cases where it's desirable for MM to think that a given host
always has an up-to-date copy of an entire category.  These are rare,
but do exist.  This field is false by default, and requires someone in
group sysadmin to change it.

Delete unused HostCategory.has_up2date_dirs().