make mirrorlist.py and mirrorlist_server work right together
mirrorlist_server now takes ~40MB RSS which is great.

mirrorlist.py unfortunately takes ~14 minutes to run.  It would take
<1m if instead of running a separate SQL query for each Directory, it
made a single big query.  However, the trim() code was really broken
and was multiplying the size of the results instead of trimming it
down, and I couldn't find the bug.  So I reverted back to 1000 SQL
queries and working code.  Room for improvement.