SCHEMA CHANGE: mirrorlist: return a longer ordered list of mirrors
For Product Versions where we know the yum prioritymethod=fallback is
set by default (such as Fedora 8 and later), we can provide additional
entries to the mirrorlist to be used if the earlier entries fail for
some reason.  Think mirrors in netblocks should return the local
mirror, and a geoip-based mirror in case the local mirror is down
briefly.

This does not affect lookups done to non-repository directories using
path=... because we don't store the Version associated with that dir
in the database at all.