report_mirror checkin: don't create HostCategoryDir entries for unknown Dirs
WPI.edu reported that the mirrorlist for some directories was not
returning them when they were marked up2date.  The root cause is that
_uploaded_config() would create a HostCategoryDir with a
directory=None foreign key, when given a directory from the client
that wasn't known about on the master servers.  (This happens with
.~tmp~ dirs too, which is just plain bloat).

So, fix it to not create HostCategoryDir() objects for unknown
directories.

And, I need to nuke all the HCDs that have directory=None at some
point to clear them out.