SCHEMA CHANGE: add Directory.readable
During pre-bitflip days, content is staged on the master mirrors, and
is available for download to the mirrors, but the directories have
permissions such that world users may not read/execute into the new
directories.  Then on release day, the "bitflip" occurs, in which the
directory permissions are made world read/executable.

Mirrors checking in using report_mirror will record they have content
in the unreadable directories.  This is good, we want this.

The crawler however, being a "world" user, will be unable to read
content from those directories, and will mark them as not up-to-date,
which we don't want.

This patch introduces Directory.readable, which is generally true,
except during pre-bitflip days.  The crawler must ignore unreadable
directories.  The update-master-directory-list script must notice when
the bitflip has occurred, and change the state of those directories in
the database.