SCHEMA CHANGE: add DirectoryExclusiveHost table, use it in mirrorlist
Some directories are so special that we only want specific "trusted"
mirrors to serve that content.  Such directory/host mappings can be
specified in this table.  This is then used in the big mirrorlist
query to throw out hosts not listed in this table for directories that
are.

(merge of same commit on master branch)