server/mirrormanager: replaced a few hardcoded "Fedora"'s with a config option
In order to get mirrormanager running for rpmfusion.org I replaced a few
hardcoded "Fedora"'s with a config option. This is only the start but
the first few pages look already pretty good.
I am not sure what to do about the copyright statement. I have no idea
how is the copyright owner of the code and if that has to appear on the
each page. For rpmfusion it will be probably wrong if all is copyrighted
be Red Hat.