mirrorlist-server: added option to analyse log for other days than current day
Add option -o (--offset) which can be used to specify
the number of days which should be used to subtract
from today's date. This is then the date which will
be used for logfile analysis. Only entries from that
day will be used to create the pie charts and the output
filenames will also reflect that offset.