fix categorymap logic
categorymap() was explicitly listing categories for product=='Fedora'
which should be there.  But this broke when category 'Fedora Secondary
Arches' was added in the database, but not added to categorymap.  This
patch changes categorymap logic to instead delete known incorrect
categories (e.g. Fedora Extras for Fedora < 3), so newly added
categories will always present in the resulting categorymap.  If they
later need to be removed, we can edit categorymap at that time.