SQLify publiclist page generation; display arm arch on publiclist chooser
publiclist generation had been running a big sql query, then throwing
away most of the results, then using SQLObject to retrieve most of the
same data to put in the publiclist.kid form.  This seemed wasteful to
me.  The publiclist generation for the top-level index page was
different than the per-product/version/arch pages too, which just
seems wrong.

So, the sql query for publiclist_hosts() was expanded to return the
information needed by the publiclist.kid template.

unrelated change, this also adds arm to the list of architectures
shown in the publiclist chooser.