only display available architectures in publiclist chooser
The publiclist chooser (top right on each publiclist page) has been
displaying a growing number of primary and secondary architecures, for
all products and versions, regardless of if there are any mirrors for
that particular product/version/arch combination.

This patch changes how the publiclist chooser is created, by first
generating the full set of product/version/arch host lists when the
/publiclist/ URL is hit.  Thereon, requests for URLs below that, such
as /publiclist/Fedora/9/i386/ use the now-cached host lists, and the
publiclist chooser omits those architectures for which there are no
mirrors.