umdl: fix rsync listing parsing
On occasion, rsync will throw a message such as
  file has vanished: "/pub/fedora/linux/updates/10/x86_64/thunderbird-2.0.0.19-1.fc10.x86_64.rpm"
into the directory listing stream.  The parser wouldn't handle this
and would die.  Fix the parser to ignore malformed lines.

Also, detection of a line as being a directory was prone to failure,
if the MOTD included a line that began with the letter 'd'.  Make this
recognition somewhat more robust.