SCHEMA CHANGE: add Directory.ctime
We now track the last changed time on each directory.

In future changes, umdl will use this to skip over directories which
haven't changed, and the crawler in --canary mode will use it to not
probe directories which haven't changed in some time.  Eventually we
can publish this information so mirrors can see what files have
changed in the last X hours, to know what to sync.