model: publiclist_hosts() SQL wasn't quite right
In the always_up2date case, it was returning all hosts that had that
category, for all versions and all arches, regardless of if that
category has content matching the requested version/architecture.

This patch fixes the SQL query to take into account the directories
for each category, matching only those that also have content for the
requested architecture.