move turbogears.url() calls from controller to templates
CSRF code appends the _csrf_token to the end of all urls generated by
turbogears.url().  We had been keying off the form value action=
(specifically, the last word in the action), to know what we were
doing with the form.  Appending ?_csrf_token=... caused that to fail.

It's more appropriate to do the tg.url(action) inside the templates
anyhow.