Add lookup by Autonomous System Number
The internet uses BGP to advertise routes.  We've been using the
internet2 routes for some time.  This extends that work to look up a
client's IP address in a global BGP routing table, to determine it's
ASN.  If we have a Host on that ASN, that will serve all clients on
that ASN, then we'll put that host on the list.  This happens after
each Host's netblock list is looked up, but before the GeoIP lookups, on
the grounds that being on the same AS is the least expensive, and most
likely fastest, method for a given ISP/client pairing.  Hosts may
still use netblock listings, but if their BGP announcements are sane,
then everyone benefits automatically.

Thanks to routeviews.org for posting the BGP tables, and for their DNS
reverse lookup mechanism, which gave me the idea do to this lookup by
AS.