mirrorlist_server: store tables as ints to reduce memory usage
Storing the global_netblocks list as IPy.IP objects consumes upwards
of 230MB of RAM.  Storing the list as a tuple of ints keeps the whole
process under 100MB.  This does away with the OrderedIP and
OrderedNetblock classes, and replaces them with a list of tuples
(ipint, prefixlen, ipversion, asn) which can be natively sorted and
bisected too.  Adds lookup_ip(list, ip), where IP is an IPy.IP object,
and after finding the appropriate item in list, that item is turned
into an IP object to compare if the IP is inside a netblock.

In the end, easier to maintain code, and much less memory consumption.