mirrorlist_server: use radix tree for netblock lookups
py-radix provides the proper API for looking up netblocks, including
nested netblocks, used by routers.  The int-based list could return
the wrong answer for nested netblocks.

This adds a dependency on py-radix, and increases the memory footprint
from 116MB to 165MB for the mirrorlist_server process, but it returns
the right answer.

Thanks to Peter Poeml of the openSUSE project (maintainer of
mirrorbrain) for the advice.