ageFileDetails(): account for propogation delays
scenario: f12 update pushed 2010-04-03T05:00:31.359362.  The following
f12 update was pushed 2010-04-09T01:30:03.61679, about 6 days later.
So, when the later update got pushed, MM "forgot" about the earlier
update (because it was more than 2 days "stale"), but the mirrors
hadn't yet picked up the new update, so yum refused to use any of the
mirrors.  And of course, it happened simultaneous with 13-Beta getting
staged to the mirrors so mirrors couldn't pick up the F12 update as
quickly as they normally would.

This hadn't bitten us before, when we had a max 7 days stale rule,
because we have been pushing updates faster than every 7 days.  But
with the 2-day rule, we get bit.

This change adds a new config option,
mirrormanager.max_propogation_days (default=2), and changes the
default max_stale_days to 3.  When a new FileDetail is created, any
that are stale will get deleted, leaving the most recent, and also
leaving the second-most recent if the most recent one is newer than
max_propogation_days.  This ensures that mirrors that are pretty
current, but haven't had a chance to catch an update within a few
hours, are still usable by yum, at least until the crawler finds them
to be not up-to-date.