templates: use constant titles
After hours of pain, I give up.  I can't have variable titles where
one option is to not pass a title and have it auto-generate.  At the
time the python calls to do this magic would happen, not enough is
loaded for the python calls to succeed.  This results in odd errors
complaining that static.kid can't be read, which isn't the case at all
(it just can't be compiled).

Now all the pages have <title>MirrorManager</title> hard coded.