mirrormanager.wsgi run 4 instances per app server (up from 2)
Reports from mirror admins that report_mirror is issuing "connection
refused" messages.  This may be because we have added mirrors, but
have relatively few processes available to accept report_mirror
requests.

This change doubles the number of waiting web UI / XMLRPC processes
running on each app server, which will increase the number of
connections to the database also.  A similar change was made recently
to increase the number of mirrorlist WSGI apps running to solve a
similar problem.