mirrorlist_client: sanitize input into UTF-8
Users may put all sorts of odd URL-escaped characters onto the URLs.
When this happens, mirrorlist_server.py child thread would crash
trying to write the byte string with escaped '\x80' characters inside
it into the error report header.  When the child crashes,
mirrorlist_client.wsgi would spin forever eating 100% CPU waiting for
a response from the server that would never come.

This patch sanitizes the input query args into a byte string that only
contains UTF-8 characters.  This fixes this particular source of crash
in the server and subsequent hang of the client.  It would still be
good if the client could time out or otherwise recognize when the
server is never coming back.