mirrorlist_client: a better way to handle socket timeouts
blocking sockets, calling recv(), may block forever if the server end
doesn't send anything for some reason.  Don't let that happen.  Python
has a socket.settimeout() capability.  We'll use that to let any
individual operation (except the select()) take up to 5 seconds (they
should all be in the microsecond range, so this is very generous), and
let select() continue to wait for 60 seconds for the server to respond
at all.

If a timeout happens, an exception is raised, which is caught by the
caller and a HTTP 503 returned to the web client.