umdl: don't use move_repository_from_development()
For Fedora 13, when releases/13 was created, this routine moved the
Repository pointers from development/13 to releases/13 as expected.
However, releases/13 was not world readable yet, so all users of
mirrorlist?repo=fedora-13 broke.

I'm not sure we can do this in umdl.  We would have to know the
permissions, and hope they don't get more restrictive after creation.
Ugh.  So, nuke it from umdl for now, and consider doing this in the
releaseday script (to be added at some point).