mirrorlist_server: fix radix tree lookups
don't overload use of tree data field 'hosts' with content that isn't
a list of hosts.  Code reuse is good, but make it obvious, and there's
no point overloading the data field as it's already a dict, so named
elements are good.

extract data from netblock_country_cache correctly.  It's supposed to
be an iterable, because it's handed as result.extend(foo), while we
were putting it into a plain element, not an iterable.  Whoops.