SCHEMA CHANGE: HostPeerASN.asn NOT NULL
Require this field to be NOT NULL.  There's no reason for a host to
have an entry in this table that would be NULL.  It's OK in the Host
table of course.

I'm not going to force this in a schema update, because this table is
new in MM 1.4, and no one is using it for real yet.