umdl: fix exception catching
Failure seen in umdl when when trying to create a new Directory object
for a newly created directory not previously seen by umdl (and
therefore did not exist in the database).

Now that we're using a SQL query cache instead of querying the
database for a single item, we don't catch SQLObjectNotFound, we need
to catch KeyError.