crawler: handle zero-length files
The crawler, when doing HTTP HEAD calls to retrieve info about a file,
gets its content length from the HTTP header.  If a file is
zero-length, the HTTP server may not return a Content-Length header.
This was causing the crawler to believe the file (and thus its
directory) was not up-to-date, even though it was up-to-date.

Fix is to check that content-length value is present, and if not
present, don't rely upon it for determining up-to-date-edness.