umdl: speed up short_filelist(), keep only most recent 3 files
short_filelist() was sorting and truncating the list on each addition
to the list (once per file in a directory).  That's horribly
inefficient.  Fix that.

Keep data on only the 3 most recent files, not 10.  This should
greatly speed up crawler_perhost as it has to do checks on all the
files in these lists.  And if you do have the most recent 3, you will
have the most recent 10.  Why 3?  I'm still a little paranoid.