don't add an extra / to mirrorlist URLs for files in the topdir
Fixes https://fedorahosted.org/fedora-infrastructure/ticket/3469

This only happens on file requests to the topdir of a category,
not on file requests to files in a subdirectory. The failure is
specific to S3, as other web servers squash the second / of the pair,
but S3 does not. This thus doesn't affect downloads of the packages
from S3 (except perhaps after the initial run which downloads the
repository key from S3), but does affect downloading of the GPG keys
from the topdirs. Once the key is retrieved from _another_ server, yum
should then start using S3 for future package downloads.

The fix is to prevent adding the extraneous / in the first place.