change crawler timeout from 60 minutes to 90 minutes
The mirrormanager crawler has a 60 minute timeout - if it can't crawl
the mirror in 60 minutes, it assumes it's running slow enough not to
be able to serve traffic reasonably, and marks it as not up-to-date,
dropping it from the mirror lists.

With the new alphabetical Packages/[0-9a-z]/*.rpm directory structure,
the crawler now as an additional ~36 directories to search in each of
24 (currently) directory trees, at 10 RPMs each being searched, at ~6
RPMs/second being checked, is another 24 minutes more than was
previously needed.  This is pushing mirrors that were previously
completing in just over 30 minutes, over the timeout.

This change bumps up the default timeout from 60 minutes to 90 minutes.