add report_mirror --exclude-from= option to exclude directories
This limits the directories sent by report_mirror to the MM database.
This is useful when a mirror has directories (e.g. .tmp.XXXXXX, .snapshot)
due to the storage subsystem being used (e.g. NetApp NFS) which exist
on the mirror but are not part of the tree being mirrored.