mirrorlist.py: fix shrink()
The current shrink() routine could fail if the end value was an
integer, because then assignment of an already-recorded value would
appear to be string character re-assignment.

Fix the syntax to be explicit that we are updating the original dict
passed in, with the keys we've walked through to find the actual
value'd object, and then set the value to the already-recorded value.