model.py: fix RLIKE usage
The previous attempt to make RLIKE vs ~ database agnostic made the
mistake of using the stock RLIKE sqlrepr implementation.  This was
causing the database table.column referred to in the RLIKE statement
to instead be treated as the explicit constant string 'table.column'
which is not what the pattern should be testing against - it will
always fail.

This patch extends the stock RLIKE implementation, to not turn the
table.column reference into a constant string.

This routine is still really slow, on the order of 15 seconds per
product + regexp (version + architecture) pair.  For 18+ versions * 11
architectures, this is _really_ slow.  This routine alone could take
30-45 minutes against the Fedora tree.