mirrorlist_server: catch SIGTERM and cleanup
Presently, any signal sent to the process will most likely interrupt
ss.serve_forever() in its select() routine, which raises an error.
This then kills the whole process, which is not what we want.

What we want is for serve_forever() to get restarted if SIGHUP comes
in, and for the program to cleanly exit if SIGTERM comes in.  That's
what this patch does.