mirrorlist_server: remove _ordered_netblocks() usage
_ordered_netblocks() usage is broken.  It is taking a hostid, and
calling IP() on it, which is plain wrong.

The goal of _ordered_netblocks() is to, given a list of (hostid,
netblock), where netblock is the block that contains the client's IP
address that we're looking up, we want to later be able to sort that
list by the size of the netblock.  However, do_netblocks() isn't
returning us such a list, it's only returning the list of hostids.
So, we need to fix do_netblocks(), then we can fix
_ordered_netblocks().  We will do neither here.

This patch drops the use of _ordered_netblocks(), reverting to the
v1.3 behavior of simply randomizing the list of hosts that claim a
given client IP address in its netblocks.  We can revisit the better
fix later.