Fix sorting results by preferred netblock size
In a previous commit, we stopped trying to sort the netblock_results
list, and left it up to a random shuffle.  We instead really want to
sort this list from smallest (most precise) netblock size to largest,
so that an individual company's mirror may be preferred over that of
the company's upstream ISP.  This patch fixes the return of the data
from the tree_lookup() function to recognize the node data may be a
list or a singleton and handle accordingly.