mirrorlist_server: reduce memory usage considerably
Rather than loading in the entire 500k entries of the global netblock
list, only load in the ~30k entries that match ASNs that our mirrors
have explicitly claimed as being relevant to them, either as the
Host.asn or in the HostPeerASN list.

This has the effect of reducing memory size by about half for the
overall program, which, should improve fork() time and reduce failure
to insufficient RAM available.