fix run_rsync() in umdl and crawler
The crawler was failing to properly mark up-to-date or not when using
rsync.  It turns out that the rsync results from a fairly complete
mirror can be nearly 100MB itself, which was sitting in memory (and
could cause memory problems).  So, let's write that to a safe
temporary file.

Fixing this requires a corresponding change in umdl.

And as tempfile.SpooledTemporaryFile only appeared in python 2.6, and
coindentally that's what RHEL6 has, fix up the spec to require python
2.6 or above.

Furthermore, the crawler was setting all_files outside a loop, instead
of _inside_ the loop where it belonged.  That's a bug.  And it was
later testing all_files inside a per-file loop, instead of after all
those files had been processed - wasteful.