set mirrorlist WSGI processes=45, threads=1
We had what appears to be a lockup of mod_wsgi / apache on one of the
Fedora mirrorlist servers overnight.  It may be related to this change
in MM 1.4, where we had set it to 6 processes, 25 threads each.  It's
worth a try to change it back to what FI has been running in
production since 2010.