fix move-devel-to-release
In the upgrade from MM 1.3. to 1.4 in Fedora Infrastructure, it was
noted that we could not make the Repository() objects simply "move"
from development/ to releases/ as we thought, because there were
already Repository objects on the releases/ directories, but with
incorrect data.  We needed to update those, and zero out fields on the
ones in development/.  That's what this patch does.

In MM 1.4, it should be that Repository objects won't get created on
releases/ directories.  So we'll have to revisit this about when
Fedora 20 branches from rawhide to be sure it works the way we want.