crawler: handle TryLater better
We weren't exiting the per-directory test loop when we got a
TryLater.  So we weren't really "trying later".  Also, we were
canceling the timer but continuing to scan (considered bad form...).

Now, add a larger and larger delay, up to 64 seconds, per test, when
given a TryLater.  and don't cancel the timer.  Eventually the timer
will time out and kill us, that's fine.  At least we won't be pounding
the remote server the whole time.