crawler: fix os.waitpid(), capture more debugging
The first argument to os.waitpid() was 0, meaning wait until _any_
process in our process group (in our case, the rsync calls by
crawler_perhost) would complete.  That's not what we want!  We only
care about our own children (the crawler_perhost process itself)
completing.  This bug was causing all invocations using rsync to fail,
and of course, only when run under the master crawler, not when
crawler_perhost was run stand-alone, which made it hard to track
down.  Ugh.

To help with the tracking, start capturing crawler_perhost stdout and
stderr.

Also be a little more verbose when removing a completed child from the
active_children list.