SCHEMA UPDATE: add Host.last_crawled_duration, use it in crawler
We want to start the crawlers in decending time order of the last
crawl duration.  This lets us start crawlers that will take a long
time first, and crawlers that will complete quickly, later.  In this
order, the overall completion time should be the fastest possible, as
a long-running job won't get stuck after an arbitrarily long bunch of
faster jobs.

Set the last_crawled_duration in the crawler after crawler_perhost has
completed, and use the value in crawler to sort the list of hosts to
be crawled.