SCHEMA CHANGE: last_crawl_duration is now BigIntCol
First, I misnamed this field in the previous commit message.  Ignore
that.

Next, last_crawl_duration is now a BigIntCol with default=0, which
contains the duration in seconds.  This can be obtained using
timedelta.total_seconds() on python => 2.7, and we fake it on lower
versions.  This is good enough for our purposes.