mv-instrepos: fix for rawhide (again)
So, there are already Repository pointers for the directories that
fedora-install-rawhide would have pointed to.  Adding the new
Repository pointers causes the cache pickle to blow up (need to figure
out why...).

Instead, we need a RepositoryRedirect in here,
fromRepo=u'fedora-install-rawhide', toRepo=u'rawhide').  And use the
unique constraint in the database to be sure we don't get several.