model: speed up publiclist_hosts() greatly
This query could take minutes to execute, as it was essentially
joining the entire database together across all the critical
relations.  It was doing a naked join on both host_category_dir (so
the results had all host_category_dirs tested, only to throw away all
that data), as well as a naked join on host_category +
category_directory + directory.

I realized we could use a subquery, LIMIT 1,  with EXISTS to handle
those naked joins cleanly.  That's what this does.

Queries that used to take minutes now take ~10 seconds, and queries
thereafter on the same or similar queries now take 0.1s, much
improved.

This will speed up generate-publiclist a bit, and reduce load on the
database.