fix mirrorlist pickle generation
it was failing to look up a parent directory, which is odd, as the parent
should have been seen first.  in reality, we don't care about
the ordered_mirrorlist setting for much of anything, especially
non-repository directories, so, just nuke that lookup.

move the parent_dir() function to lib.py.