Adapt code to new MM2 layout.
The functionality to disable/enable a host via report_mirror
used to be part of the model.Host class. It has moved to a
standalone function and therefore can no longer access the
model.Host fields via self.. The functionality to disable/enable
host has also been moved to an earlier step where the host
data is modified and not the HostCategory data.