Auto disable hosts which have a URL configured but which does not exist
There are a few hosts in the DB which do not have the configured URL
anymore. Auto disable the hosts if the broken URL persists for a
certain number of crawls. Before any other access to the host a
connection is made if either via http or ftp the host can be reached.