Scan the directories which are supposed to be on each mirror.
The crawler was only looking at the files which a mirror actually
had and not at the files it is supposed to have. This lead to the
situation that a newly added host or category was never crawled
successfully. Now the crawler checks the list of files which are
supposed to be on the host and compares it with the actual content
of the crawled mirror.