Add option to scan only one category
As the time required to scan all categories is increasing this
offers the option to only scan one category. With the parameter

  --category="Fedora EPEL"

it is possible to scan only certain category. The additional option
'--list' gives an overview of the available categories. This opens
up the possibility to start category specific scan upon fedmsg
notifications and not run a scan over the whole tree every two hours.