Trying explicit garbage collection
The crawler requires a lot of memory. With the explicit garbage
collection at the end of each crawler thread the crawler has not
been OOM killed for the last 4 runs. If this actually helps is
not very clear but it seems to slightly decrease the OOM killed
frequency. Unfortunately it seems difficult to tell which part
of the crawler and which thread is using the memory and for what.