Fix version handling
The wrong version was retrieved from the database and thus the new
repositories were pointing to some bogus version. Now first the
category is retrieved from the database and then with the category
and the product name the correct version is retrieved from the
database.

Additionally the parameter handling was optimized and wrong comments
were removed.