Make timeout global
In preparation of the 'soon' coming signal handler the timeout
is no longer passed around as a parameter but is changed to
global variable. Especially with threads this is not the nicest
design but it provides an infrastructure to easily signal all threads
that they should abort as soon as possible.