Decrease httplib timeout
In an earlier commit (2ce0c54acff7ff7a6cae8a0f5dfd0bdd2568a321) the
possibility to add a timeout value for httplib/ftplib was added. The
same timeout value for connection establishment was used as for the
crawling of all categories. So it was probably better than waiting for
ever but it still does not make much sense as according to the
documentation

 If the optional timeout parameter is given, blocking operations
 (like connection attempts) will timeout after that many seconds
 (if it is not given, the global default timeout setting is used).

Let's try it with a much short value of 1 minutes. If during this time
it was not possible to open a HTTP or FTP connection something is not
right with the mirror.