Add signal handler
This adds a signal handler which offers the possibility to 'gracefully'
shut down the crawler with SIGARLM. Once the crawler receives SIGARLM it
reduces the global timeout to -1, which should abort any crawl. Once all
threads are informed about the situation (shut down) the main thread
waits 5 more minutes for all threads to finish. A HTTP or FTP crawl
should stop during that period, only rsync crawls will not abort as we
do not have (yet) a way to intercept rsync runs.
If after 5 minutes some threads do not have finished an additional
minute is waited to retrieve the results. If some threads are still
running at this time all threads are forcefully terminated.