Add a propagation test option to the crawler
Running the crawler with --propgation will check all active mirrors
and print out the checksum of a repomd.xml file. The default check
is using 'development/rawhide/x86_64/os/repodata'. With the option
--proppath this can be changed to any other path providing a repomd.xml
file. The output of this propagation check looks like:

<logheader>:URL::<check-url>::SHA256::<checksum downloaded>::<date>::<<checksum in DB>

A later commit will provide a script to create diagrams from this output.

Signed-off-by: Adrian Reber <adrian@lisas.de>