Deleting URLs if site or host is removed.
Deleting a mirror site and/or host did not correctly delete the URLs:

https://github.com/fedora-infra/mirrormanager2/issues/115
https://fedorahosted.org/mirrormanager/ticket/61

As URLs can only exist once in the database the URLs with the deleted
site and/or hosts are still in the database and cannot be added to
another host or deleted. Right now these left-over URLs have to be
deleted manually.

Signed-off-by: Adrian Reber <adrian@lisas.de>